Floating header: move embed-only logic from shared.tsx into embed.tsx
Address chihsuan's review feedback that wp-admin chrome handling (h1
suppression, Screen Options / Help proxy icons, chrome-only narrow bar)
only ever applies on classic admin pages, so the orchestration belongs in
EmbedHeader rather than the BaseHeader shared between embed and React
routes.

- New `use-wp-admin-chrome.ts` custom hook owns the DOM detection
  (`hasH1`, `hasScreenOptions`, `hasContextualHelp`), the
  `activeMetaIcon` state synced via MutationObserver, the
  cross-system click handler that closes wp-admin dropdowns when an
  activity-panel tab is clicked, and the `triggerMetaIcon` function
  with its one-shot observer chain for gear ↔ help mutual exclusion.
- `BaseHeader` shrinks to a dumb layout component. New props
  `suppressTitle`, `compact`, and `trailingItems` let the caller drive
  embed-specific behavior; non-embed `Header` just doesn't pass them and
  gets the default (always render title, full height, no trailing items).
- `EmbedHeader` calls the hook, composes the gear / ? icons into
  `trailingItems`, and passes `suppressTitle` / `compact` based on
  whether wp-admin rendered its own h1.

Behavior is unchanged across Home, Settings, Edit Order, Edit Product,
Add Product, Products list. The seam between embed and non-embed is now
explicit in the prop surface rather than a runtime `isEmbedded` branch
inside the shared component.
